requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-secret-manager==2.15.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==2.10
lxml==4.9.2
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
requests==2.25.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
